requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
numpy==2.2.1
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
